fix: integer in chart

+++ lib/pages/Home/ChartOnHome.dart
@@ -258,7 +258,7 @@ class BarChartSample1State extends State<BarChartSample1> {
               ),
               children: <TextSpan>[
                 TextSpan(
-                  text: (rod.toY - 1).toString(),
+                  text: (rod.toY - 1).toInt().toString(),
                   style: const TextStyle(
                     color: Colors.yellow,
                     fontSize: 16,
